Migrate Take tests to munit
